Update imports and fix typos in multiple files
This commit updates the import paths to relative imports in `ng_tools.py` for consistency across the module. It also fixes a typo in the method name `get_resoltuion` to `get_resolution` and corrects a variable reference error in `ng_resolution.py` for accurate default settings. Additionally, it moves `NODE_DISPLAY_NAME_MAPPINGS` to be imported alongside `NODE_CLASS_MAPPINGS` in `__init__.py`.